+++ packages/skills/CHANGELOG.md
@@ -1,5 +1,97 @@
 # @agent-native/skills
 
+## 0.2.337
+
+### Patch Changes
+
+- Updated dependencies [76a9cca]
+- Updated dependencies [76a9cca]
+- Updated dependencies [76a9cca]
+- Updated dependencies [76a9cca]
+  - @agent-native/core@0.102.2
+
+## 0.2.336
+
+### Patch Changes
+
+- Updated dependencies [ab33ae7]
+  - @agent-native/core@0.102.1
+
+## 0.2.335
+
+### Patch Changes
+
+- Updated dependencies [6498492]
+  - @agent-native/core@0.102.0
+
+## 0.2.334
+
+### Patch Changes
+
+- Updated dependencies [866d8f0]
+- Updated dependencies [866d8f0]
+  - @agent-native/core@0.101.14
+
+## 0.2.333
+
+### Patch Changes
+
+- Updated dependencies [e503433]
+- Updated dependencies [e503433]
+- Updated dependencies [e503433]
+  - @agent-native/core@0.101.13
+
+## 0.2.332
+
+### Patch Changes
+
+- Updated dependencies [2ff004e]
+  - @agent-native/core@0.101.12
+
+## 0.2.331
+
+### Patch Changes
+
+- Updated dependencies [fd0ddb8]
+  - @agent-native/core@0.101.11
+
+## 0.2.330
+
+### Patch Changes
+
+- Updated dependencies [3453dd3]
+  - @agent-native/core@0.101.10
+
+## 0.2.329
+
+### Patch Changes
+
+- Updated dependencies [bc29c82]
+  - @agent-native/core@0.101.9
+
+## 0.2.328
+
+### Patch Changes
+
+- Updated dependencies [b023dad]
+- Updated dependencies [b023dad]
+- Updated dependencies [b023dad]
+  - @agent-native/core@0.101.8
+
+## 0.2.327
+
+### Patch Changes
+
+- Updated dependencies [70caa6f]
+  - @agent-native/core@0.101.7
+
+## 0.2.326
+
+### Patch Changes
+
+- Updated dependencies [dd47e0a]
+  - @agent-native/core@0.101.6
+
 ## 0.2.325
 
 ### Patch Changes